Update linter to use unconvert

+++ oci/stats.go
@@ -54,7 +54,7 @@ func getMemLimit(cgroupLimit uint64) uint64 {
 		return cgroupLimit
 	}
 
-	physicalLimit := uint64(si.Totalram)
+	physicalLimit := si.Totalram
 	if cgroupLimit > physicalLimit {
 		return physicalLimit
 	}
